moved get sequence
